Graph: Update css.
This is where the #map style is changed, not the previous commit. The #graph element also has a margin-right of 20px.
The PCRS has some style attribute somewhere trhat is forcing the scroll bar to overlap the graph. http://cs.toronto.edu/~iansb/PCRS-Graph/graph_viz.html does not have this problem.