Graph: Nelson Milk bug fixed (desciption in description).
doable nodes with 2 text nodes for the label of the node would not rearrange properly because the counter rects text was being inserted between the first label text node and the second.
In an example with "Nelson" and "Milk" as text nodes, "Milk" would do wacky things. This has been fixed. Hooray.