Restructured the schema to include test suite code in the problems_java_problem table.
The test case table rows will automatically be generated in PCRS java when
you save a problem. This commit breaks the submission page, but the
problem creation page is mostly functional. On the problem creation
page, it has the option to edit test cases, but this should be blocked
in the future since they are automatically generated.