Updated the "test suite has changed" hook.
Before, if anything in the test suite changed (including descriptions
and comments), you were not allowed to save your changes. Now, it will
strip superflous information like comments and whitespace before
comparing the old and new test suite code.

Note, however, that there are a couple limitations:
- Comments after code won't be stripped out (e.g. "} // end")
- Assertion messages can't be changed without validation errors

Assertion message parsing will be more complicated, and require a lot of
testing. For instance, what if the instructor has the code:
assertEquals(add(2, 2), 4);

A naive approach would be to remove the first assert argument if there
are 3 arguments. However, in the above example this would not be
desired, as it would result in "assertEquals( 2), 4);".