Fixed the grade table and clear submissions form-reloading.
The grade table hid the description when the window was too small. It
will now always be shown.
Clicking "Reset Submissions" in the dialog will now reload the page
along with any form data. So if a user clicks 'Save & Attempt', then
clears the submissions, it will save and go to the attempt page.