Student code insertion bug fix & refactoring
When we introduced the [file] tag, it caused a lot of issues with
inserting other code tags. One of which was how the problem_form
automatically inserts student code tags. This issue is now fixed.
However, the insert tag buttons are still broken. Once we start splitting
the starter code into multiple files, it will be much easier to fix
them, so I didn't put too much effort there.