Fixed some issue with code tag highlighting and blocking.
In this commit, I completely ripped out the code tag blocking code.
Personally, I found it extremely annoying when I manually changed a
code tag and it would block me. In place of this guard, I
updated the highlighting code to indicate (in red) if the user placed mismatching
code tags.

If this is an issue, let me know