Fixed a bug with JS compression.
Both the bugs Diane and Jen mentioned in their emails should be fixed
now. The issue was an object (Handlebars) being accessed before it was
created.

This bug seems to only occur when compression is enabled, which is why I
didn't run into it earlier. In the future, I will test my commits with
compression enabled so this doesn't happen again.