Started adding multiple tabs to the Java editor.
You can add multiple files currently, but the visualizer won't display
them.

Also note for future reference: I am going to try to retain
functionality for all of the languages supported on pythontutor.com.
In the future it might be a good idea to start using this single pytutor port.