Re-added breakpoints to the visualizer.
When I updated the visualizer to support multiple files, it broke the
breakpoint functionality. It is now fixed!

I noticed some funky behavior when jumping to breakpoints. For instance,
once in a while pressing "Forward" will go to the next non-breakpoint
step, but other times it will jump all the way to the next breakpoint.
This was not introduced recently - there is the same behavior on the main
PythonTutor project.