Fixed the disappearing objects issue.
Currently, it will _NOT_ hide objects as they are returning (as
discussed in todays meeting).

Right now when objects return, they will appear on the heap with no
apparent references to them. The next step, a reference will be
assigned.

To indicate that an object has been returned but not assigned, I was thinking
about putting a "returned value" tag on the heap object.

To try the changes you will have to "ant build" JavaJail again