Updated JavaJail.
The new version has more verbose error reporting for missing
requirements in student code.  Specifically, it is more verbose about
the class signature of the first file submitted (it must be a public class),
and the main method of said class.

"ant bulid" in languages/java/execution/JavaJail is required