Fixed the issue Jen was running into yesterday.
It turned out to be an issue with the code editor. In certain
situations, such as if you pasted code in certain places or pressed
"Shift+Enter" at the top of a file, code would be inserted outside both
student_code and blocked sections, so it would be ignored. This, of course,
would cause unexpected compile errors.