Added a "Load Solution" button.
Copy-pasting while testing solutions was _really_ tedious for
instructors before. Now there is a "Load Solution" button that does it
for you!

Note, however, that submissions may explode if the code tags in the
solution box do not coorespond to the code tags in the starter code.

This won't affect students, but instructors should keep that in mind.